PKG -- [types] Abstract number-to-string conversion into its own file.
